+++ R/convert_notes.R
@@ -45,6 +45,7 @@
 
 convert_notes <- function(d, code = NULL, anchors = NULL, nThread = parallel::detectCores()-1) {
   .SD=.N=.I=.GRP=.BY=.EACHI=..=..cols=.SDcols=i=j=time_to_db=..which_ids_to=..which_ids_from=..collapse <- NULL
+  options(future.globals.maxSize = +Inf)
 
   message(paste0("Extracting information from notes free text."))
 
@@ -100,6 +101,8 @@ convert_notes <- function(d, code = NULL, anchors = NULL, nThread = parallel::de
   colnames(result) <- cols
 
   result <- cbind(d, result)
+
+  on.exit(options(future.globals.maxSize = 1.0 * 1e9))
   future::plan(future::sequential)
   return(result)
 }