+++ R/convert_rfv.R
@@ -33,6 +33,7 @@ convert_rfv <- function(d, code = "rfv_concept_id", codes_to_find = NULL,
                         collapse = NULL, code_time = "time_rfv_start", aggr_type = "earliest", nThread = parallel::detectCores()-1) {
 
   .SD=.N=.I=.GRP=.BY=.EACHI=..=..code=.SDcols=i=j=time_to_db=..which_ids_to=..which_ids_from=..collapse=. <- NULL
+  options(future.globals.maxSize = +Inf)
 
   #Initialize multicore
   if(nThread == 1 | length(codes_to_find) == 1) {
@@ -80,6 +81,7 @@ convert_rfv <- function(d, code = "rfv_concept_id", codes_to_find = NULL,
           diag_coll
         }
       }
+    on.exit(options(future.globals.maxSize = 1.0 * 1e9))
     future::plan(future::sequential)
 
     if(is.null(collapse)) { #Remove unnecessary info and combine with original data if non-collapse
@@ -90,6 +92,7 @@ convert_rfv <- function(d, code = "rfv_concept_id", codes_to_find = NULL,
     }
     return(result)
   } else { #If no diagnoses
+    on.exit(options(future.globals.maxSize = 1.0 * 1e9))
     future::plan(future::sequential)
     return(d)
   }