+++ R/parse_ids.R
@@ -21,6 +21,8 @@
 parse_ids <- function(str, num = NULL, sep = ":", id_length = "standard", perc = 0.6, nThread = parallel::detectCores()-1) {
 
   .SD=.N=.I=.GRP=.BY=.EACHI=..=..cols=.SDcols=i=j=time_to_db=..which_ids_to=..which_ids_from <- NULL
+  options(future.globals.maxSize = +Inf)
+
 
   #Initialize multicore
   if(nThread == 1) {
@@ -68,6 +70,7 @@ parse_ids <- function(str, num = NULL, sep = ":", id_length = "standard", perc =
       df_ids
     }
 
+  on.exit(options(future.globals.maxSize = 1.0 * 1e9))
   future::plan(future::sequential)
   return(result)
 }